change class StartClass

+++ src/main/java/consoleCalculator/Addition.java
@@ -1,4 +1,4 @@
-// Класс, в котором описаны ввод чисел, сложение и вывод результата в консоль
+// Класс сложения
 
 package consoleCalculator;
 
@@ -6,13 +6,8 @@ public class Addition implements Operations{
     public void add() {
 
         DataInput dataInput = new DataInput();
-        //Вызов метода ввода с клавиатуры
         dataInput.inputValue();
-
-        //обращаемся к интерфейсу, где прописана логика
-        int resultAdding = additionInterface(dataInput.valueClass.getFirstValue(), dataInput.valueClass.getSecondValue());
-
-        //вывод результата суммы
+        double resultAdding = additionInterface(dataInput.valueClass.getFirstValue(), dataInput.valueClass.getSecondValue());
         System.out.println("------------------------- " +
                 "\nAddition result: " + resultAdding);
     }
